fix: rename Tooth to SwapIf
